+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Warning for all Users when cookies are currently disabled.
+- Prompt for all Users to use the latest version of Chrome or Firefox.
 
 ## [1.0.2] - 2019-01-10